Refactor TilePointerIndex
