+++ src/openrct2/libopenrct2.vcxproj
@@ -494,6 +494,7 @@
     <ClInclude Include="world\TileElement.h" />
     <ClInclude Include="world\TileElementsView.h" />
     <ClInclude Include="world\TileInspector.h" />
+    <ClInclude Include="world\TilePointerIndex.hpp" />
     <ClInclude Include="world\Wall.h" />
     <ClInclude Include="world\Water.h" />
   </ItemGroup>
